+++ src/style-spec/diff.test.ts
@@ -256,6 +256,8 @@ test('diff', () => {
         {command: 'setPitch', args: [1]}
     ]);
 
+
+
     expect(diffStyles({
         light: {
             anchor: 'map',
@@ -408,6 +410,42 @@ test('diff', () => {
         transition: 'transition'
     })).toEqual([
         {command: 'setTransition', args: ['transition']}
+    ]);
+
+    expect(diffStyles({
+        sprite: 'a'
+    }, {
+        sprite: [{id:'default', url:'a'}]
+    })).toEqual([]);
+
+    expect(diffStyles({
+        sprite: 'a'
+    }, {
+        sprite: 'b'
+    })).toEqual([
+        {command: 'setSprite', args: ['b']},
+    ]);
+
+    expect(diffStyles({
+        sprite: 'a'
+    }, {
+        sprite: [{'id':'default', 'url':'b'}]
+    })).toEqual([
+        {command: 'setSprite', args: [{'id':'default','url':'b'}]},
+    ]);
+
+    expect(diffStyles({
+        glyphs: 'a'
+    }, {
+        glyphs: 'a'
+    })).toEqual([]);
+
+    expect(diffStyles({
+        glyphs: 'a'
+    }, {
+        glyphs: 'b'
+    })).toEqual([
+        {command: 'setGlyphs', args: ['b']},
     ]);
 
 });